python/2019-11-22_Content_Aggregator_Bot/requirements.txt
discord.py
praw
pafy
youtube_dl
wikipedia
google
beautifulsoup4
requests